SPIFFS init (test)
The functions were rebuilt and the project structure was changed.
