+++ components/test/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "test.c"
+                       INCLUDE_DIRS "."
+                       REQUIRES main uspiffs)
