+++ components/test/test.h
@@ -1,0 +1,8 @@
+#ifndef __TEST_H__
+#define __TEST_H__
+
+#include <uspiffs.h>
+
+void test_main();
+
+#endif
