+++ main/main.h
@@ -8,4 +8,6 @@
 
 #include <uspiffs.h>
 
+#include <test.h>
+
 #endif
